chore: 添加 .gitignore 并修复 flutter analyze 报错
- 创建 .gitignore（排除 build/、.dart_tool/、native/build/ 等）
- 修复 native_bridge.dart: _getFrame 字段类型 Int64→int 及调用签名
- 删除 video_info.dart 中未使用的 video_file.dart 导入
- 删除 preview_panel.dart 中未使用的 dart:typed_data 导入
- 删除 convert_provider.dart 中未使用的 dart:async 导入

+++ lib/ffi/native_bridge.dart
@@ -19,7 +19,7 @@ class NativeBridge {
   late final int Function(
     Pointer<Void>,
     Pointer<Uint8>,
-    Int64,
+    int,
     Pointer<Int32>,
     Pointer<Int32>,
   ) _getFrame;
